Update Card Border Radius and Adjust Margins for Consistency
Adjusted the border radius of cards in Masaa.dart and Sabah.dart from 4 to 12 for topLeft and topRight corners to enhance visual consistency across pages. Additionally, modified the margin settings in azkar.dart to include a bottom margin of 72 for the specific card at index 131, ensuring better spacing and alignment within the layout. Removed an unnecessary SizedBox with a height of 20 from the end of a column in azkar.dart to streamline the UI and improve code readability.